feat: readme, fix crypto/utils crates
